Moved checkSurroundings function to utility file
